+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsstrengthsbasedneedsassessmentsapi/oasys/controller/response/OasysAssessmentVersionResponse.kt
@@ -7,7 +7,7 @@ import java.util.UUID
 
 data class OasysAssessmentVersionResponse(
   val sanAssessmentId: UUID,
-  val sanAssessmentVersion: Long,
+  val sanAssessmentVersion: Int,
   val sanAssessmentData: AssessmentResponse,
   val lastUpdatedTimestamp: LocalDateTime,
 ) {
